Expose peer address in RPC contexts
Make the remote peer's address available to client and server
interceptors, and to server handlers, by annoting the respective request
contexts with peer.Peer values. The same peer.Peer value is reused for
all requests on a connection to reduce GC pressure. This does run the
risk, however, of the value getting mutated in one request affecting the
value read back in other requests on the same connection.

Signed-off-by: Cory Snider <csnider@mirantis.com>