added the ttrpc stress utility
This change adds a stress utility which can be used for stress testing
the ttrpc connection. This tool represents a simple client-server
interaction where the client sends continuous requests to the server,
and the server responds with the same data, allowing for testing of
concurrent request handling and response verification.

This tool is adapted from https://github.com/kevpar/ttrpcstress.git
which was written by Kevin Parsons.

Signed-off-by: Harsh Rawat <harshrawat@microsoft.com>